refactor: reorganize config
