+++ packages/core/src/cli/init.ts
@@ -1,7 +1,7 @@
 import path from 'node:path';
-import { loadConfig as baseLoadConfig } from '../config';
 import { createRsbuild } from '../createRsbuild';
 import { castArray, ensureAbsolutePath } from '../helpers';
+import { loadConfig as baseLoadConfig } from '../loadConfig';
 import { logger } from '../logger';
 import { watchFilesForRestart } from '../restart';
 import type { RsbuildInstance } from '../types';
